DMStag

The following DMStag routines are available:

PETSc.DMStag — Method

DMStag(
    petsclib::PetscLib
    comm::MPI.Comm,
    boundary_type::NTuple{D, DMBoundaryType},
    global_dim::NTuple{D, Integer},
    dof_per_node::NTuple{1 + D, Integer},
    stencil_width::Integer,
    stencil_type = DMSTAG_STENCIL_BOX;
    points_per_proc::Tuple,
    processors::Tuple,
    setfromoptions = true,
    dmsetup = true,
    options...
)

Creates a D-dimensional distributed staggered array with the options specified using keyword arguments.

The Tuple dof_per_node specifies how many degrees of freedom are at all the staggerings in the order:

1D: (vertex, element)

2D: (vertex, edge, element)

3D: (vertex, edge, face, element)

If keyword argument points_per_proc[k] isa Vector{petsclib.PetscInt} then this specifies the points per processor in dimension k.

If keyword argument processors[k] isa Integer then this specifies the number of processors used in dimension k; ignored when D == 1.

If keyword argument setfromoptions == true then setfromoptions! called.

If keyword argument dmsetup == true then setup! is called.

When D == 1 the stencil_type argument is not required and ignored if specified.

External Links

PETSc Manual: DMSTAG/DMStagCreate1d

PETSc Manual: DMSTAG/DMStagCreate2d

PETSc Manual: DMSTAG/DMStagCreate3d

PETSc.boundarytypes — Method

boundarytypes(dm::AbstractDMStag)

Returns an NTuple{3, DMBoundaryType} with boundary types for the given staggered dm. on

External Links

PETSc Manual: DMSTAG/DMStagGetBoundaryTypes

PETSc.getcorners — Method

getcorners(dm::AbstractDMStag)

Returns a NamedTuple with the global indices (excluding ghost points) of the lower and upper corners as well as the size. Also included is nextra of the number of extra partial elements in each direction.

External Links

PETSc Manual: DMDA/DMStagGetCorners

PETSc.getghostcorners — Method

getghostcorners(dm::AbstractDMStag)

Returns a NamedTuple with the global indices (including ghost points) of the lower and upper corners as well as the size.

External Links

PETSc Manual: DMDA/DMDAGetGhostCorners

PETSc.globalsize — Method

globalsize(dm::AbstractDMStag)

return an NTuple{3, PetscInt} of the global size of the staggered dm.

Only the first getdimension(dm) values are set; trailing values set to 1.

External Links

PETSc Manual: DMStag/DMStagGetGlobalSizes

PETSc.isfirstrank — Method

isfirstrank(dm::AbstractDMStag)

Returns an NTuple{3, Bool} of whether this MPI rank is first rank in each dimension of the dm

External Links

PETSc Manual: DMDA/DMStagGetIsFirstRank

PETSc.islastrank — Method

islastrank(dm::AbstractDMStag)

Returns an NTuple{3, Bool} of whether this MPI rank is last rank in each dimension of the dm

Only the first getdimension(dm) values are set; trailing values set to DM_BOUNDARY_NONE.

External Links

PETSc Manual: DMDA/DMStagGetIsLastRank

PETSc.localsize — Method

localsize(dm::AbstractDMStag)

return an NTuple{3, PetscInt} of the loval size of the staggered dm on this MPI rank.

Only the first getdimension(dm) values are set; trailing values set to 1.

External Links

PETSc Manual: DMStag/DMStagGetLocalSizes