KSP

The following KSP routines are available:

PETSc.KSP — Method

KSP(A::AbstractMat, P::AbstractMat{PetscLib} = A; options...)

Create a KSP using the matrix A and preconditioner construction matrix P with the options.

The communicator is obtained from A and if it has size 1 then the garbage collector is set, otherwise the user is responsible for calling destroy.

External Links

PETSc Manual: KSP/KSPCreate

PETSc Manual: KSP/KSPSetOperators

PETSc Manual: KSP/KSPSetFromOptions

PETSc.createvecs — Method

createvecs(ksp::AbstractKSP; nright = 0, nleft = 0)

Create nright right and nleft left vectors compatible with the ksp. Returned object V has Tuple members V.right and V.left containing the vectors.

External Links

PETSc Manual: KSP/KSPCreateVecs