KSP

The following KSP routines are available:

PETSc.KSP — Method

KSP([petsclib,] A::SparseMatrixCSC; options...)

Create a KSP with the sparse matrix A using the petsclib. If petsclib is not given, the default library will be used`.

PETSc.KSP — Method

KSP(da::AbstractDM; options...)

Construct a PETSc Krylov subspace solver from the distributed mesh

Any PETSc options prefixed with ksp_ and pc_ can be passed as keywords.

External Links

PETSc Manual: KSP/KSPCreate

PETSc Manual: KSP/KSPSetDM

PETSc Manual: KSP/KSPSetFromOptions

PETSc.KSP — Method

KSP(A::AbstractMat, P::AbstractMat{PetscLib} = A; options...)

Create a KSP using the matrix A and preconditioner construction matrix P with the options.

The communicator is obtained from A and if it has size 1 then the garbage collector is set, otherwise the user is responsible for calling destroy.

External Links

PETSc Manual: KSP/KSPCreate

PETSc Manual: KSP/KSPSetOperators

PETSc Manual: KSP/KSPSetFromOptions

PETSc.KSPPtr — Type

KSPPtr(petsclib, ksp::CKS, own)

Container type for a PETSc KSP that is just a raw pointer.

PETSc.createvecs — Method

createvecs(ksp::AbstractKSP; nright = 0, nleft = 0)

Create nright right and nleft left vectors compatible with the ksp. Returned object V has Tuple members V.right and V.left containing the vectors.

External Links

PETSc Manual: KSP/KSPCreateVecs

PETSc.getDMDA — Method

getDMDA(ksp::AbstractKSP)

Get dmda for ksp

The returned dmda is owned by the ksp

External Links

PETSc Manual: KSP/KSPGetDM

PETSc.getsolution — Method

getsolution(ksp)

returns the solution vector stored in the ksp. This function does not make a new copy of the vector, it merely returns the vector stored in the ksp

External Links

PETSc Manual: KSP/KSPGetSolution

PETSc.setcomputeoperators! — Method

setcomputeoperators!(ksp::AbstractKSP, ops!::Function)
setcomputeoperators!(ops!::Function, ksp::AbstractKSP)

Define ops! to be the compute operators function for the ksp. A call to ops!(A, P, new_ksp) should set the elements of the PETSc matrix linear operator A and preconditioning matrix P based on the new_ksp.

Note

The new_ksp passed to ops! may not be the same as the ksp passed to setcomputeoperators!.

External Links

PETSc Manual: KSP/KSPSetComputeOperators

PETSc.setcomputerhs! — Method

setcomputerhs!(ksp::AbstractKSP, rhs!::Function)
setcomputerhs!(rhs!::Function, ksp::AbstractKSP)

Define rhs! to be the right-hand side function of the ksp. A call to rhs!(b, new_ksp) should set the elements of the PETSc vector b based on the new_ksp.

Note

The new_ksp passed to rhs! may not be the same as the ksp passed to setcomputerhs!.

External Links

PETSc Manual: KSP/KSPSetComputeRHS