KSP

The following KSP routines are available:

PETSc.Fn_KSPComputeOperators — Type

struct Fn_KSPComputeOperators{T} end

Type used to wrap ComputeOperators! functions in KSP

PETSc.Fn_KSPComputeRHS — Type

struct Fn_KSPComputeRHS{T} end

Type used to wrap ComputeRHS! functions in KSP

PETSc.KSP — Method

KSP(da::AbstractDM; options...)

Construct a PETSc Krylov subspace solver from the distributed mesh

Any PETSc options prefixed with ksp_ and pc_ can be passed as keywords.

see PETSc manual

PETSc.KSP — Method

KSP(A, P; options...)

Construct a PETSc Krylov subspace solver.

Any PETSc options prefixed with ksp_ and pc_ can be passed as keywords.

PETSc.KSPSetComputeOperators! — Function

KSPSetComputeOperators!(
    ksp::KSP{Number},
    ComputeOperators!,
    ctx = C_NULL,
)

Set the linear operators function ComputeOperators! for the ksp using the user ctx. ComputeOperators! should be callable with four arguments of type (::KSP{Number}, ::Mat, ::Mat, ::Ptr); see PETSc manual

PETSc.KSPSetComputeRHS! — Function

KSPSetComputeRHS!(
    ksp::KSP{Number},
    ComputeRHS!,
    ctx = C_NULL,
)

Set the right-hand side function ComputeRHS! for the ksp using the user ctx. ComputeRHS! should be callable with three arguments of type (::KSP{Number}, ::Vec, ::Ptr); see PETSc manual

PETSc.iters — Function

iters(ksp::KSP)

Gets the current iteration number; if the solve! is complete, returns the number of iterations used.

External Links

PETSc Manual: KSP/KSPGetIterationNumber

PETSc.resnorm — Function

resnorm(ksp::KSP)

Gets the last (approximate preconditioned) residual norm that has been computed.

External Links

PETSc Manual: KSP/KSPGetResidualNorm