List of all functions

Here a summary of all functions:

PETSc.AbstractOptions

PETSc.DMDA

PETSc.DMGlobalVec

PETSc.DMLocalVec

PETSc.GlobalOptions

PETSc.KSP

PETSc.KSP

PETSc.KSP

PETSc.KSPPtr

PETSc.MatAIJ

PETSc.MatAIJ

PETSc.MatPtr

PETSc.MatSeqAIJ

PETSc.MatSeqDense

PETSc.MatShell

PETSc.Options

PETSc.SNES

PETSc.SNESPtr

PETSc.VecGhost

PETSc.VecMPI

PETSc.VecPtr

PETSc.VecSeq

PETSc.VecSeqWithArray

Base.empty

PETSc.assemble!

PETSc.assemble!

PETSc.assemblybegin!

PETSc.assemblybegin!

PETSc.assemblyend!

PETSc.assemblyend!

PETSc.coordinatesDMLocalVec

PETSc.createvecs

PETSc.createvecs

PETSc.finalize

PETSc.finalized

PETSc.getDMDA

PETSc.getDMDA

PETSc.getcoordinateDM

PETSc.getcorners

PETSc.getdimension

PETSc.getghostcorners

PETSc.getinfo

PETSc.getlocalcoordinatearray

PETSc.getlocalform

PETSc.getpetsctype

PETSc.getpetsctype

PETSc.getsolution

PETSc.gettype

PETSc.getvalues!

PETSc.getvalues!

PETSc.ghostupdatebegin!

PETSc.ghostupdateend!

PETSc.initialize

PETSc.initialized

PETSc.localinteriorlinearindex

PETSc.ndofs

PETSc.ownershiprange

PETSc.ownershiprange

PETSc.parse_options

PETSc.reshapelocalarray

PETSc.restorelocalform!

PETSc.setDM!

PETSc.setcomputeoperators!

PETSc.setcomputerhs!

PETSc.setfromoptions!

PETSc.setfunction!

PETSc.setjacobian!

PETSc.setuniformcoordinates!

PETSc.setup!

PETSc.setup!

PETSc.setvalues!

PETSc.setvalues!

PETSc.setvalues!

PETSc.typedget

PETSc.unsafe_localarray

PETSc.update!

PETSc.update!

PETSc.updatebegin!

PETSc.updatebegin!

PETSc.updateend!

PETSc.updateend!

PETSc.view

PETSc.with

PETSc.withlocalarray!

PETSc.withlocalform

PETSc.AbstractOptions — Type

AbstractOptions{PetscLib <: PetscLibType}

Abstract type of PETSc solver options.

PETSc.DMDA — Method

DMDA(
    petsclib::PetscLib
    comm::MPI.Comm,
    boundary_type::NTuple{D, DMBoundaryType},
    global_dim::NTuple{D, Integer},
    dof_per_node::Integer,
    stencil_width::Integer,
    stencil_type;
    points_per_proc::Tuple,
    processors::Tuple,
    setfromoptions = true,
    dmsetup = true,
    options...
)

Creates a D-dimensional distributed array with the options specified using keyword arguments.

If keyword argument points_per_proc[k] isa Vector{petsclib.PetscInt} then this specifies the points per processor in dimension k.

If keyword argument processors[k] isa Integer then this specifies the number of processors used in dimension k; ignored when D == 1.

If keyword argument setfromoptions == true then setfromoptions! called.

If keyword argument dmsetup == true then setup! is called.

When D == 1 the stencil_type argument is not required and ignored if specified.

External Links

PETSc Manual: DMDA/DMDACreate1d

PETSc Manual: DMDA/DMDACreate2d

PETSc Manual: DMDA/DMDACreate3d

PETSc.DMGlobalVec — Type

DMGlobalVec(dm::AbstractDM)

returns a global vector from the dm object.

External Links

PETSc Manual: DM/DMCreateGlobalVector

PETSc.DMLocalVec — Type

DMLocalVec(dm::AbstractDM)

returns a local vector from the dm object (has space for ghost).

External Links

PETSc Manual: DM/DMCreateLocalVector

PETSc.GlobalOptions — Type

GlobalOptions{PetscLib <: PetscLibType}

The PETSc global options database.

PETSc.MatAIJ — Type

MatAIJ(
    petsclib::PetscLib,
    comm::MPI.Comm,
    loc_num_rows::Integer,
    loc_num_cols::Integer,
    diag_nonzeros::Union{Integer, Vector},
    off_diag_nonzeros::Union{Integer, Vector};
    glo_num_rows = PETSC_DETERMINE,
    glo_num_cols = PETSC_DETERMINE,
    setup = true
) where {PetscLib <: PetscLibType}

Create an MPI PETSc sparse array on the comm using AIJ format (also known as a compressed sparse row or CSR format) of size glo_num_rows X glo_num_cols with local size loc_num_rows X loc_num_cols.

The diagonal block and off-diagonal block non-zeros are diag_nonzeros and off_diag_nonzeros which can be either an integer (same for all rows) or a Vector of PetscInts with on entry per row.

Memory allocation is handled by PETSc and garbage collection can be used.

If glo_num_rows isa Integer or glo_num_cols isa Integer then the corresponding local variable can be PETSC_DECIDE.

If setup == true then setup! is called

External Links

PETSc Manual: Mat/MatCreateAIJ

PETSc Manual: Mat/MatSetUp

Note

The user is responsible for calling destroy(mat) on the MatAIJ since this cannot be handled by the garbage collector do to the MPI nature of the object.

PETSc.MatAIJ — Method

MatAIJ(dm::AbstractDM)

Generates a matrix from the dm object.

External Links

PETSc Manual: DM/DMCreateMatrix

PETSc.MatPtr — Type

MatPtr(petsclib, mat::CMat)

Container type for a PETSc Mat that is just a raw pointer.

PETSc.MatSeqAIJ — Type

MatSeqAIJ(petsclib, num_rows, num_cols, nonzeros)

Create a PETSc serial sparse array using AIJ format (also known as a compressed sparse row or CSR format) of size num_rows X num_cols with nonzeros per row

If nonzeros is an Integer the same number of non-zeros will be used for each row, if nonzeros is a Vector{PetscInt} then one value must be specified for each row.

Memory allocation is handled by PETSc and garbage collection can be used.

External Links

PETSc Manual: Mat/MatCreateSeqAIJ

PETSc.MatSeqDense — Type

MatSeqDense{PetscLib, PetscScalar}

PETSc dense array. This wraps a Julia Matrix{PetscScalar} object.

External Links

PETSc Manual: Mat/MatCreateSeqDense

PETSc.MatShell — Type

MatShell(
    petsclib::PetscLib,
    obj::OType,
    comm::MPI.Comm,
    local_rows,
    local_cols,
    global_rows = LibPETSc.PETSC_DECIDE,
    global_cols = LibPETSc.PETSC_DECIDE,
)

Create a global_rows X global_cols PETSc shell matrix object wrapping obj with local size local_rows X local_cols.

The obj will be registered as an MATOP_MULT function and if if obj is a Function, then the multiply action obj(y,x); otherwise it calls mul!(y, obj, x).

if comm == MPI.COMM_SELF then the garbage connector can finalize the object, otherwise the user is responsible for calling destroy.

External Links

PETSc Manual: Mat/MatCreateShell

PETSc Manual: Mat/MatShellSetOperation

PETSc Manual: Mat/MATOP_MULT

PETSc.Options — Type

Options{PetscLib <: PetscLibType}(kw -> arg, ...)
Options(petsclib, kw -> arg, ...)

Create a PETSc options data structure for the petsclib.

For construction a set of keyword argment pairs should be given. If the option has no value it should be set to nothing or true. Setting an option to false will cause the option not to be set on the PETSc options table.

Examples

julia> using PETSc

julia> petsclib = PETSc.petsclibs[1];

julia> PETSc.initialize(petsclib)

julia> opt = PETSc.Options(
                         petsclib,
                         ksp_monitor = nothing,
                         ksp_view = true,
                         pc_type = "mg",
                         pc_mg_levels = 1,
                         false_opt = false,
                     )
#PETSc Option Table entries:
-ksp_monitor
-ksp_view
-pc_mg_levels 1
-pc_type mg
#End of PETSc Option Table entries

julia> opt["ksp_monitor"]
""

julia> opt["pc_type"]
"mg"

julia> opt["pc_type"] = "ilu"
"ilu"

julia> opt["pc_type"]
"ilu"

julia> opt["false_opt"]
ERROR: KeyError: key "bad_key" not found

julia> opt["bad_key"]
ERROR: KeyError: key "bad_key" not found

External Links

PETSc Manual: Sys/PetscOptionsCreate

PETSc.SNES — Method

SNES(
    petsclib::PetscLib
    comm::MPI.Comm;
    setfromoptions = true,
    options...
)

Creates a PETSc nonlinear solve context on the communicator comm

If keyword argument setfromoptions == true then setfromoptions! called on the created SNES context.

External Links

PETSc Manual: SNES/SNESCreate

PETSc Manual: SNES/SNESSetFromOptions

PETSc.SNESPtr — Type

SNESPtr{PetscLib}(ptr, age)

Container type for a PETSc SNES that is just a raw pointer.

PETSc.VecGhost — Type

VecGhost(
     petsclib,
     comm:MPI.Comm,
     local_length
     ghost::Vector{PetscInt};
     global_length = PETSC_DETERMINE,
     num_ghost = length(ghost),
)

An sequentially-stored MPI PETSc vector for petsclib.PetscScalar of local length local_length and global length global_length with ghost elements.

If global_length isa Int then local_length can be set to PETSC_DECIDE.

External Links

PETSc Manual: Vec/VecCreateGhost

Note

The user is responsible for calling destroy(vec) on the Vec since this cannot be handled by the garbage collector do to the MPI nature of the object.

PETSc.VecMPI — Type

VecMPI(
     petsclib,
     comm:MPI.Comm,
     local_length;
     global_length = PETSC_DETERMINE
)

An sequentially-stored MPI PETSc vector for petsclib.PetscScalar of local length local_length and global length global_length without ghost elements.

If global_length isa Int then local_length can be set to PETSC_DECIDE in which case PETSc will decide the local_length.

External Links

PETSc Manual: Vec/VecCreateMPI

Note

The user is responsible for calling destroy(vec) on the Vec since this cannot be handled by the garbage collector do to the MPI nature of the object.

PETSc.VecPtr — Type

VecPtr(petsclib, v::CVec, own)

Container type for a PETSc Vec that is just a raw pointer.

If the own then the finalizer is set on the vector; calling destroy when !own is a no-op.

PETSc.VecSeq — Type

VecSeq(petsclib, n::Int)

A standard, sequentially-stored serial PETSc vector for petsclib.PetscScalar of length n.

External Links

PETSc Manual: Vec/VecCreateSeq

PETSc.VecSeqWithArray — Type

VecSeqWithArray(petsclib, v::Vector)

A standard, sequentially-stored serial PETSc vector, wrapping the Julia vector v.

This reuses the array v as storage, and so v should not be resize!-ed or otherwise have its length modified while the PETSc object exists.

This should only be need to be called for more advanced uses, for most simple usecases, users should be able to pass Vectors directly and have the wrapping performed automatically

External Links

PETSc Manual: Vec/VecCreateSeqWithArray

Base.empty — Method

similar(da::DMDA)

return an uninitialized DMDA struct.

PETSc.assemble! — Function

assemble!(M::AbstractMat[, t::MatAssemblyType = MAT_FINAL_ASSEMBLY)

Assemble the matrix M with assembly type t.

For overlapping assembly see assemblybegin! and assemblyend!

External Links

PETSc Manual: Mat/MatAssemblyBegin

PETSc Manual: Mat/MatAssemblyEnd

PETSc.assemble! — Method

assemble!(v::AbstractVec)

Assemble the vector v.

For overlapping assembly see assemblybegin! and assemblyend!

External Links

PETSc Manual: Vec/VecAssemblyBegin

PETSc Manual: Vec/VecAssemblyEnd

PETSc.assemblybegin! — Method

assemblybegin!(M::AbstractMat[, t::MatAssemblyType = MAT_FINAL_ASSEMBLY)

Begin assembly of the matrix M with assembly type t; finished with assemblyend!.

External Links

PETSc Manual: Mat/MatAssemblyBegin

PETSc.assemblybegin! — Method

assemblybegin!(vec::AbstractVec)

Begin assembling vec

External Links

PETSc Manual: Vec/VecAssemblyBegin

PETSc.assemblyend! — Method

assemblyend!(M::AbstractMat[, t::MatAssemblyType = MAT_FINAL_ASSEMBLY)

Finish assembly of the matrix M with assembly type t; start assembly with assemblybegin!.

External Links

PETSc Manual: Mat/MatAssemblyEnd

PETSc.assemblyend! — Method

assemblyend!(vec::AbstractVec)

Finish assembling vec

External Links

PETSc Manual: Vec/VecAssemblyEnd

PETSc.coordinatesDMLocalVec — Method

coordinatesDMLocalVec(dm::AbstractDM)

Gets a local vector with the coordinates associated with dm.

Note that the returned vector is borrowed from the dm and is not a new vector.

External Links

PETSc Manual: DM/DMGetCoordinatesLocal

PETSc.createvecs — Method

createvecs(
    M::AbstractMat{PetscLib},
)

Returns vectors V which are compatible with M. A right compatible vectors is V.right and a left compatible vector is V.left; positionally these are returned as (right, left)

The created vectors are not garbage collected and should be destroyed with destroy.

External Links

PETSc Manual: Mat/MatCreateVecs

PETSc.finalize — Method

finalize(petsclib)

Finalize the petsclib, if no petsclib is given then all PETSc.petsclibs will be finalized.

External Links

PETSc Manual: Sys/PetscFinalize

PETSc.finalized — Method

finalized(petsclib)

Check if petsclib is finalized

External Links

PETSc Manual: Sys/PetscFinalized

PETSc.getDMDA — Method

getDMDA(snes::AbstractSNES)

Get dmda for snes

The returned dmda is owned by the snes

External Links

PETSc Manual: SNES/SNESGetDM

PETSc.getcoordinateDM — Method

getcoordinateDM(dm::AbstractDM)

Create a coord_dm for the coordinates of dm.

External Links

PETSc Manual: DM/DMGetCoordinateDM

PETSc.getcorners — Method

getcorners(da::AbstractDMDA)

Returns a NamedTuple with the global indices (excluding ghost points) of the lower and upper corners as well as the size.

External Links

PETSc Manual: DMDA/DMDAGetCorners

PETSc.getdimension — Method

getdimension(dm::AbstractDM)

Return the topological dimension of the dm

External Links

PETSc Manual: DM/DMGetDimension

PETSc.getghostcorners — Method

getghostcorners(da::AbstractDMDA)

Returns a NamedTuple with the global indices (including ghost points) of the lower and upper corners as well as the size.

External Links

PETSc Manual: DMDA/DMDAGetGhostCorners

PETSc.getinfo — Method

getinfo(da::AbstractDMDA)

Get the info associated with the distributed array da. Returns V which has fields

dim

global_size (Tuple of length 3)

procs_per_dim (Tuple of length 3)

dof_per_node

boundary_type (Tuple of length 3)

stencil_width

stencil_type

External Links

PETSc Manual: DMDA/DMDAGetInfo

PETSc.getlocalcoordinatearray — Method

getlocalcoordinatearray(da::AbstractDMDA)

Returns a NamedTuple with OffsetArrays that contain the local coordinates and that can be addressed uisng global indices

PETSc.getlocalform — Method

getlocalform(vec::AbstractVec)

Obtains the local ghosted representation of a Vec.

Note

When done with the object the user should call restorelocalform!

External Links

PETSc Manual: Vec/VecGhostGetLocalForm

PETSc.getpetsctype — Method

getpetsctype(mat::AbstractMat)

return a string with the matrix type

External Links

PETSc Manual: Mat/MatGetType

PETSc.getpetsctype — Method

getpetsctype(vec::AbstractVec)

return a string with the vector type

External Links

PETSc Manual: Vec/VecGetType

PETSc.gettype — Method

gettype(dm::AbstractDM)

Gets type name of the dm

External Links

PETSc Manual: DM/DMGetType

PETSc.getvalues! — Method

getvalues!(
    rowvals::Array{PetscScalar};
    M::AbstractMat{PetscLib},
    row0idxs::Vector{PetscInt},
    col0idxs::Vector{PetscInt},
    num_rows = length(row0idxs),
    num_cols = length(col0idxs)
)

get values of the matrix M with base-0 row and column indices row0idxs and col0idxs inserting the values rowvals.

If the keyword arguments num_rows or num_cols is specified then only the first num_rows * num_cols values of rowvals will be used.

External Links

PETSc Manual: Mat/MatGetValues

PETSc.getvalues! — Method

getvalues!(
    vals::Array{PetscScalar},
    v::AbstractVec,
    indices::Vector{PetscInt},
)

Get the 0-based global indices of vec into the preallocated array vals.

Warning

This function uses 0-based indexing!

External Links

PETSc Manual: Vec/VecGetValues

PETSc.ghostupdatebegin! — Method

ghostupdatebegin!(
    vec::AbstractVec,
    insertmode = INSERT_VALUES,
    scattermode = SCATTER_FORWARD,
)

Begins scattering vec to the local or global representations

External Links

PETSc Manual: Vec/VecGhostUpdateBegin

PETSc.ghostupdateend! — Method

ghostupdateend!(
    vec::AbstractVec,
    insertmode = INSERT_VALUES,
    scattermode = SCATTER_FORWARD,
)

Finishes scattering vec to the local or global representations

External Links

PETSc Manual: Vec/VecGhostUpdateEnd

PETSc.initialize — Method

initialize([petsclib])

Initialized the petsclib, if no petsclib is given then all PETSc.petsclibs will be initialized.

Additionally:

This will initialize MPI if it has not already been initialized.

It will disable the PETSc signal handler (via Sys/PetscPopSignalHandler

Add an atexit hook to call PETSc.finalize.

External Links

PETSc Manual: Sys/PetscInitializeNoArguments

PETSc.initialized — Method

initialized(petsclib)

Check if petsclib is initialized

External Links

PETSc Manual: Sys/PetscInitialized

PETSc.localinteriorlinearindex — Method

localinteriorlinearindex(dmda::AbstractDMDA)

returns the linear indices associated with the degrees of freedom own by this MPI rank embedded in the ghost index space for the dmda

PETSc.ndofs — Method

ndofs(da::AbstractDMDA)

Return the number of dofs in for da

External Links

PETSc Manual: DMDA/DMDAGetDof

PETSc.ownershiprange — Method

ownershiprange(mat::AbstractMat, [base_one = true])

The range of row indices owned by this processor, assuming that the mat is laid out with the first n1 rows on the first processor, next n2 rows on the second, etc. For certain parallel layouts this range may not be well defined.

If the optional argument base_one == true then base-1 indexing is used, otherwise base-0 index is used.

Note

unlike the C function, the range returned is inclusive (idx_first:idx_last)

External Links

PETSc Manual: Mat/MatGetOwnershipRange

PETSc.ownershiprange — Method

ownershiprange(vec::AbstractVec, [base_one = true])

The range of indices owned by this processor, assuming that the vec is laid out with the first n1 elements on the first processor, next n2 elements on the second, etc. For certain parallel layouts this range may not be well defined.

If the optional argument base_one == true then base-1 indexing is used, otherwise base-0 index is used.

Note

unlike the C function, the range returned is inclusive (idx_first:idx_last)

External Links

PETSc Manual: Vec/VecGetOwnershipRange

PETSc.parse_options — Method

parse_options(args::Vector{String})

Parse the args vector into a NamedTuple that can be used as the options for the PETSc solvers.

julia --project file.jl -ksp_monitor -pc_type mg -ksp_view -da_refine=1

PETSc.reshapelocalarray — Method

reshapelocalarray(Arr, da::AbstractDMDA{PetscLib}[, ndof = ndofs(da)])

Returns an array with the same data as Arr but reshaped as an array that can be addressed with global indexing.

PETSc.restorelocalform! — Method

restorelocalform!(local_vec::LocalVec)

Restore the local_vec to the associated global vector after a call to getlocalform.

External Links

PETSc Manual: Vec/VecGhostRestoreLocalForm

PETSc.setDM! — Method

setDM!(snes::AbstractSNES, dm::AbstractDM)

Set dm for snes

External Links

PETSc Manual: SNES/SNESSetDM

PETSc.setfromoptions! — Method

setfromoptions!(dm::DM, opts=dm.opts)

External Links

PETSc Manual: DM/DMSetFromOptions

PETSc.setfunction! — Method

setfunction!(snes::AbstractSNES, f!::Function, x::AbstractVec)
setfunction!(f!::Function, snes::AbstractSNES, x::AbstractVec)

Define f! to be the function of the snes. A call to f!(fx, snes, x) should set the elements of the PETSc vector fx based on the x.

External Links

PETSc Manual: SNES/SNESSetFunction

PETSc.setjacobian! — Function

setjacobian!(
    snes::AbstractSNES,
    updateJ!::Function,
    J::AbstractMat,
    P::AbstractMat = J
)
setjacobian!(
    updateJ!::Function,
    snes::AbstractSNES,
    J::AbstractMat,
    P::AbstractMat = J
)

Define updateJ! to be the function that updates the Jacobian of the snes.

If J == P then a call to updateJ!(J, snes, x) should set the elements of the PETSc Jacobian (approximation).

If J ≠ P then a call to updateJ!(J, P, snes, x) should set the elements of the PETSc Jacobian (approximation) and preconditioning matrix P.

External Links

PETSc Manual: SNES/SNESSetJacobian

PETSc.setuniformcoordinates! — Method

setuniformcoordinates!(
    da::DMDA
    xyzmin::NTuple{N, Real},
    xyzmax::NTuple{N, Real},
) where {N}

Set uniform coordinates for the da using the lower and upper corners defined by the NTuples xyzmin and xyzmax. If N is less than the dimension of the da then the value of the trailing coordinates is set to 0.

External Links

PETSc Manual: DMDA/DMDASetUniformCoordinates

PETSc.setup! — Method

setup!(dm::DM, opts=dm.opts)

External Links

PETSc Manual: DM/DMSetUp

PETSc.setup! — Method

setup!(mat::AbstractMat)

Set up the interal data for mat

External Links

PETSc Manual: Mat/MatSetUp

PETSc.setvalues! — Method

setvalues!(
    M::AbstractMat{PetscLib},
    row0idxs::Vector{MatStencil{PetscInt}},
    col0idxs::Vector{MatStencil{PetscInt}},
    rowvals::Array{PetscScalar},
    insertmode::InsertMode = INSERT_VALUES;
    num_rows = length(row0idxs),
    num_cols = length(col0idxs)
)

Set values of the matrix M with base-0 row and column indices row0idxs and col0idxs inserting the values rowvals.

If the keyword arguments num_rows or num_cols is specified then only the first num_rows * num_cols values of rowvals will be used.

External Links

PETSc Manual: Mat/MatSetValuesStencil

PETSc.setvalues! — Method

setvalues!(
    M::AbstractMat{PetscLib},
    row0idxs::Vector{PetscInt},
    col0idxs::Vector{PetscInt},
    rowvals::Array{PetscScalar},
    insertmode::InsertMode = INSERT_VALUES;
    num_rows = length(row0idxs),
    num_cols = length(col0idxs)
)

Set values of the matrix M with base-0 row and column indices row0idxs and col0idxs inserting the values rowvals.

If the keyword arguments num_rows or num_cols is specified then only the first num_rows * num_cols values of rowvals will be used.

External Links

PETSc Manual: Mat/MatSetValues

PETSc.setvalues! — Method

setvalues!(
    v::AbstractVec,
    indices::Vector{PetscInt},
    vals::Array{PetscScalar},
    insertmode::InsertMode,
)

Assign the values vals in 0-based global indices of vec. The insertmode can be INSERT_VALUES or ADD_VALUES.

Warning

This function uses 0-based indexing!

External Links

PETSc Manual: Vec/VecSetValues

PETSc.typedget — Method

typedget(opt::NamedTuple, key::Symbol, default::T)

Parse opt similar to get but ensures that the returned value is the same type as the default value. When T <: NTuple keys that result in a single value will be filled into an NTuple of the same length as T; in the case of strings it is parsed using split with comma delimiter

Examples

julia> opt = (tup = (1, 2, 3), string_tup = "1,2,3", string_int = "4", int = 4)
(tup = (1, 2, 3), string_tup = "1,2,3", string_int = "4", int = 4)

julia> typedget(opt, :int, 7)
4

julia> typedget(opt, :bad_key, 7)
7

julia> typedget(opt, :tup, (1, 1, 1))
(1, 2, 3)

julia> typedget(opt, :string_tup, (1, 1, 1))
tokens = SubString{String}["1", "2", "3"]
(1, 2, 3)

julia> typedget(opt, :string_int, (1, 1, 1))
tokens = SubString{String}["4"]
(4, 4, 4)

julia> typedget(opt, :int, (1, 1, 1))
(4, 4, 4)

julia> typedget(opt, :int, (1., 1., 1.))
(4.0, 4.0, 4.0)

PETSc.unsafe_localarray — Method

unsafe_localarray(vec::AbstractVec; read=true, write=true)

Return an Array{PetscScalar} containing local portion of the PETSc vec

Use read=false if the array is write-only; write=false if read-only.

Note

Base.finalize should be called on the Array before the data can be used.

External Links

PETSc Manual: Vec/VecGetArray

PETSc Manual: Vec/VecGetArrayWrite

PETSc Manual: Vec/VecGetArrayRead

PETSc Manual: Vec/VecRestoreArray

PETSc Manual: Vec/VecRestoreArrayWrite

PETSc Manual: Vec/VecRestoreArrayRead

PETSc.update! — Method

update!(
    global_vec::AbstractVec,
    local_vec::DMLocalVec,
    mode::InsertMode,
)

Updates global_vec from the local_vec with insert mode.

Communication and computation can be overlapped with updatebegin! and updateend!

External Links

PETSc Manual: DM/DMLocalToGlobal

PETSc Manual: DM/DMLocalToGlobalBegin

PETSc Manual: DM/DMLocalToGlobalEnd

PETSc.update! — Method

update!(
    local_vec::DMLocalVec,
    global_vec::AbstractVec,
    mode::InsertMode,
)

Updates local_vec from the global_vec with insert mode.

Communication and computation can be overlapped with updatebegin! and updateend!

External Links

PETSc Manual: DM/DMGlobalToLocal

PETSc Manual: DM/DMGlobalToLocalBegin

PETSc Manual: DM/DMGlobalToLocalEnd

PETSc.updatebegin! — Method

updatebegin!(
    global_vec::AbstractVec,
    local_vec::DMLocalVec,
    mode::InsertMode,
)

Begin update of global_vec from the local_vec with insert mode.

External Links

PETSc Manual: DM/DMLocalToGlobalBegin

PETSc.updatebegin! — Method

updatebegin!(
    local_vec::DMLocalVec,
    global_vec::AbstractVec,
    mode::InsertMode,
)

Begin update of local_vec from the global_vec with insert mode.

External Links

PETSc Manual: DM/DMGlobalToLocalBegin

PETSc.updateend! — Method

updateend!(
    global_vec::AbstractVec,
    local_vec::DMLocalVec,
    mode::InsertMode,
)

End update of global_vec from the local_vec with insert mode.

External Links

PETSc Manual: DM/DMLocalToGlobalEnd

PETSc.updateend! — Method

updateend!(
    local_vec::DMLocalVec,
    global_vec::AbstractVec,
    mode::InsertMode,
)

End update of local_vec from the global_vec with insert mode.

External Links

PETSc Manual: DM/DMGlobalToLocalEnd

PETSc.view — Method

view(dm::AbstractDM, [viewer])

view a dm with viewer

External Links

PETSc Manual: DM/DMView

PETSc.with — Method

with(f, opts::Options)

Call f() with the Options opts set temporarily (in addition to any global options).

PETSc.withlocalarray! — Method

withlocalarray!(
    f!,
    vecs::NTuple{N, AbstractVec};
    read::Union{Bool, NTuple{N, Bool}} = true,
    write::Union{Bool, NTuple{N, Bool}} = true,
)

Convert x to an Array{PetscScalar} using unsafe_localarray and apply the function f!.

Use read=false if the array is write-only; write=false if read-only.

Examples

```julia-repl julia> withlocalarray!(x; write=true) do x @. x .*= 2 end

julia> withlocalarray!( x, y; read = (false, true), write = (true, false) ) do x, y @. x .= 2 .+ y end

Note

Base.finalize is automatically called on the array.

PETSc.withlocalform — Method

withlocalform(f::Function, vec::AbstractVec)

Convert vec to a LocalVec and apply the function f!.

julia> withlocalform(vec) do l_vec
   # Do something with l_vec
end

Note

This wrapper handles the calling of restorelocalform! before returning.