List of all functions

Here a summary of all functions:

PETSc.AbstractOptions

PETSc.AbstractViewer

PETSc.GlobalOptions

PETSc.KSP

PETSc.MatSeqAIJ

PETSc.MatSeqDense

PETSc.MatShell

PETSc.Options

PETSc.PetscLibType

PETSc.VecSeq

PETSc.ViewerStdout

PETSc.Initialize

PETSc.Initialized

PETSc.decref

PETSc.finalize

PETSc.finalized

PETSc.incref

PETSc.inttype

PETSc.iters

PETSc.nrefs

PETSc.ownershiprange

PETSc.parse_options

PETSc.realtype

PETSc.resnorm

PETSc.scalartype

PETSc.unsafe_localarray

PETSc.with

PETSc.AbstractOptions — Type

AbstractOptions{PetscLib <: PetscLibType}

Abstract type of PETSc solver options.

PETSc.AbstractViewer — Type

AbstractViewer{PetscLib <: PetscLibType}

Abstract type of PETSc viewer.

External Links

PETSc Manual: Viewer/PetscViewer

PETSc.GlobalOptions — Type

GlobalOptions{PetscLib <: PetscLibType}

The PETSc global options database.

PETSc.MatSeqAIJ — Type

MatSeqAIJ{T}

PETSc sparse array using AIJ format (also known as a compressed sparse row or CSR format).

Memory allocation is handled by PETSc.

PETSc.MatSeqDense — Type

MatSeqDense{T}

PETSc dense array. This wraps a Julia Matrix{T} object.

PETSc.MatShell — Type

MatShell{T}(obj, m, n)

Create a m×n PETSc shell matrix object wrapping obj.

If obj is a Function, then the multiply action obj(y,x); otherwise it calls mul!(y, obj, x). This can be changed by defining PETSc._mul!.

PETSc.Options — Type

Options{PetscLib <: PetscLibType}(kw -> arg, ...)
Options(petsclib, kw -> arg, ...)

Create a PETSc options data structure for the petsclib.

For construction a set of keyword argment pairs should be given. If the option has no value it should be set to nothing or true. Setting an option to false will cause the option not to be set on the PETSc options table.

Examples

julia> using PETSc

julia> petsclib = PETSc.petsclibs[1];

julia> PETSc.initialize(petsclib)

julia> opt = PETSc.Options(
                         petsclib,
                         ksp_monitor = nothing,
                         ksp_view = true,
                         pc_type = "mg",
                         pc_mg_levels = 1,
                         false_opt = false,
                     )
#PETSc Option Table entries:
-ksp_monitor
-ksp_view
-pc_mg_levels 1
-pc_type mg
#End of PETSc Option Table entries

julia> opt["ksp_monitor"]
""

julia> opt["pc_type"]
"mg"

julia> opt["pc_type"] = "ilu"
"ilu"

julia> opt["pc_type"]
"ilu"

julia> opt["false_opt"]
ERROR: KeyError: key "bad_key" not found

julia> opt["bad_key"]
ERROR: KeyError: key "bad_key" not found

External Links

PETSc Manual: Sys/PetscOptionsCreate

PETSc.PetscLibType — Type

PetscLibType{PetscScalar, PetscInt}(petsc_library)

A container for specific PETSc libraries.

All other containers for PETSc objects should be typed on this to ensure that dispatch is correct.

PETSc.VecSeq — Type

VecSeq(v::Vector)

A standard, sequentially-stored serial PETSc vector, wrapping the Julia vector v.

This reuses the array v as storage, and so v should not be resize!-ed or otherwise have its length modified while the PETSc object exists.

This should only be need to be called for more advanced uses, for most simple usecases, users should be able to pass Vectors directly and have the wrapping performed automatically

External Links

PETSc Manual: Vec/VecCreateSeqWithArray

PETSc.ViewerStdout — Type

ViewerStdout(petsclib, comm = MPI.COMM_SELF)

Create an ASCII PetscViewer for the comm

External Links

PETSc Manual: Viewer/PETSC_VIEWER_STDOUT_

PETSc.Initialize — Function

initialize([petsclib])

Initialized the petsclib, if no petsclib is given then all PETSc.petsclibs will be initialized.

Additionally:

This will initialize MPI if it has not already been initialized.

It will disable the PETSc signal handler (via Sys/PetscPopSignalHandler

Add an atexit hook to call PETSc.finalize.

External Links

PETSc Manual: Sys/PetscInitializeNoArguments

PETSc.Initialized — Function

initialized(petsclib)

Check if petsclib is initialized

External Links

PETSc Manual: Sys/PetscInitialized

PETSc.decref — Method

decref(obj)

Decrement the reference counter for obj.

In general we don't need to use this, as we can call destroy instead.

PETSc.finalize — Function

finalize(petsclib)

Finalize the petsclib, if no petsclib is given then all PETSc.petsclibs will be finalized.

External Links

PETSc Manual: Sys/PetscFinalize

PETSc.finalized — Function

finalized(petsclib)

Check if petsclib is finalized

External Links

PETSc Manual: Sys/PetscFinalized

PETSc.incref — Method

incref(obj)

Increment the reference counter fo obj. This usually only needs to be called when accessing objects owned by other objects, e.g. via KSPGetPC.

PETSc.inttype — Method

inttype(petsclib::PetscLibType)

return the int type for the associated petsclib

PETSc.nrefs — Method

nrefs(obj)

The current reference count for obj.

PETSc.ownershiprange — Function

ownership_range(vec::AbstractVec)

The range of indices owned by this processor, assuming that the vectors are laid out with the first n1 elements on the first processor, next n2 elements on the second, etc. For certain parallel layouts this range may not be well defined.

Note: unlike the C function, the range returned is inclusive (idx_first:idx_last)

External Links

PETSc Manual: Vec/VecGetOwnershipRange

PETSc.parse_options — Method

parse_options(args::Vector{String})

Parse the args vector into a NamedTuple that can be used as the options for the PETSc solvers.

julia --project file.jl -ksp_monitor -pc_type mg -ksp_view

PETSc.realtype — Method

realtype(petsclib::PetscLibType)

return the real type for the associated petsclib

PETSc.scalartype — Method

scalartype(petsclib::PetscLibType)

return the scalar type for the associated petsclib

PETSc.unsafe_localarray — Function

unsafe_localarray(PetscScalar, ptr:CVec; read=true, write=true)

Return an Array{PetscScalar} containing local portion of the PETSc data.

finalize should be called on the Array before the data can be used.

Use read=false if the array is write-only; write=false if read-only.

PETSc.with — Method

with(f, opts::Options)

Call f() with the Options opts set temporarily (in addition to any global options).